requirements.txt
Pillow
discord.py
fuzzywuzzy[speedup]
git+https://github.com/Rapptz/asqlite@master
